+++ src/components/domains/posts/PostsFooter/index.tsx
@@ -1,0 +1,3 @@
+export const PostsFooter = () => {
+  return null;
+};
