+++ src/components/domains/posts/PostsSection/postsSection.module.scss
@@ -1,14 +1,15 @@
-.postsSection {
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  gap: 20px;
-}
+.postsSectionContainer {
+  padding: 0 8%;
 
-.postsFilterWrap {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  height: 100px;
-  gap: 10px;
+  .postsSection {
+    display: flex;
+    justify-content: center;
+    flex-wrap: wrap;
+    padding: 0 20px;
+    gap: 20px;
+  }
+
+  .postsTriggerWrap {
+    padding: 10vh 0;
+  }
 }
