+++ src/pages/PostsPage/PostsPage.scss
@@ -1,5 +1,6 @@
 .body_section {
   max-width: 1450px;
-  width: 80%;
+  padding: 0 30px;
+  width: 85%;
   margin: 0 auto;
 }
