+++ src/pages/PostsPage/index.tsx
@@ -1,11 +1,16 @@
+import { PostsFooter } from '@components/domains/posts/PostsFooter';
+import { PostsHeader } from '@components/domains/posts/PostsHeader';
 import { PostsSection } from '@components/domains/posts/PostsSection';
 import './PostsPage.scss';
 
 const PostsPage = () => {
   return (
-    <div className="body_section">
+    <main>
+      <PostsHeader />
       <PostsSection />
-    </div>
+      <PostsFooter />
+    </main>
   );
 };
+
 export default PostsPage;
